Force display of Many2Many screen after edition
Since changeset 0c2f313bf62d, the cancel of the record is not enough to refresh
the display of the edited record in the Many2Many list view. So a display call
must be triggered to clear the CellCache.

issue4391
review6721002
(grafted from c616682c4c85c68b84f1272cbab4369d16161fab)

--HG--
branch : 3.4